refactor: align CI panel composition
