+++ packages/frontend/src/components/features/agents/task-execution-ci-timestamp-line.tsx
@@ -1,0 +1,19 @@
+import type { ReactElement } from "react";
+import { humanDate } from "@/lib/task-display";
+
+export function TaskExecutionCiTimestampLine({
+  label,
+  timestamp,
+}: {
+  label: string;
+  timestamp: string;
+}): ReactElement {
+  return (
+    <span>
+      {label}{" "}
+      <time dateTime={timestamp} title={timestamp}>
+        {humanDate(timestamp)}
+      </time>
+    </span>
+  );
+}
